Move tone_hash method into module to use in both track model and repos facade
